Setup e TearDown e limpeza codigo duplicado

+++ run_robot_scripts.sh
@@ -1,28 +1,49 @@
 #!/bin/bash
 FOLDER_NAME=$1
-ROBOT_FOLDER=$1/robot
+ROBOT_FOLDER=$FOLDER_NAME/robot
 OUTPUT_FOLDER=$ROBOT_FOLDER/output
+
+ROBOT_SCRIPT_NAME=$2
 
 if [ ! -d "$FOLDER_NAME" ]; then
     printf "Please, tell a valid folder name (example: run_robot_scripts.sh organo)"
     exit 1
 fi
 
-printf "Folder selected: $FOLDER_NAME\n";
+printf "Folder selected: $FOLDER_NAME\n"
 
-printf "*****************************
+if [ ! -z "$ROBOT_SCRIPT_NAME" ]; then
+    ROBOT_SCRIPT="$ROBOT_FOLDER/$ROBOT_SCRIPT_NAME.robot"
+
+    if [ ! -f "$ROBOT_SCRIPT" ]; then
+        printf "Robot script '$ROBOT_SCRIPT' is invalid. Please omit this variable or tell a valid one"
+        exit 1
+    else
+        printf "Script selected: $ROBOT_SCRIPT\n\n"
+
+        robot --version
+        rm -rf $OUTPUT_FOLDER 2> /dev/null
+
+        printf "\n"
+        robot --outputdir $OUTPUT_FOLDER $ROBOT_SCRIPT
+    fi
+else
+    printf "\n*****************************
 * Running Robot Automations *
 *****************************\n\n"
 
-robot --version
-rm -rf $OUTPUT_FOLDER 2> /dev/null
+    robot --version
+    rm -rf $OUTPUT_FOLDER 2> /dev/null
 
-i=1
-for robot_script in $ROBOT_FOLDER/*.robot; do
-    printf "\n$i) Running $(basename $robot_script)\n"
+    i=1
+    for ROBOT_SCRIPT in $ROBOT_FOLDER/*.robot; do
+        printf "\n$i) Running $(basename $ROBOT_SCRIPT)\n"
 
-    mkdir $OUTPUT_FOLDER/script_$i 2> /dev/null
-    robot --outputdir $OUTPUT_FOLDER/script_$i $robot_script
+        mkdir $OUTPUT_FOLDER/script_$i 2> /dev/null
+        robot --outputdir $OUTPUT_FOLDER/script_$i $ROBOT_SCRIPT
 
-    i=$(( i + 1 ))
-done
+        i=$(( i + 1 ))
+    done
+fi
+
+exit 0
